+++ Changes.md
@@ -5,6 +5,7 @@ Improvements
 ------------
 
 - Viewer : Tool hotkeys now work as toggles for non-exclusive tools. This is currently noticeable on the Camera Tool - pressing <kbd>T</kbd> again will deactivate it.
+- ArnoldShader and ArnoldLight : Added activators and sections to the UI.
 
 API
 ---