+++ Changes.md
@@ -11,6 +11,7 @@ API
 ---
 
 - Added support for `gaffer.layout.activator` and `gaffer.layout.visibilityActivator` in `.mtd` files.
+- User interface sections are now hidden if all plugs in that section are hidden.
 
 0.61.9.0 (relative to 0.61.8.0)
 ========