+++ Changes.md
@@ -1,3 +1,11 @@
+0.62.x.x (relative to 0.62.0.0a2)
+========
+
+API
+---
+
+- RenderController : Added `pathForID()`, `pathsForIDs()`, `idForPath()` and `idsForPaths()` methods. These make it possible to identify an object in the scene from a `uint id` AOV.
+
 0.62.0.0a2 (relative to 0.62.0.0a1)
 ==========
 
@@ -39,7 +47,6 @@ API
 
 - ImageAlgo : Added `sortChannelNames()` function.
 - NodeAlgo : Added support for presets on compound plugs. If all child plugs have a particular preset then the parent plug is considered to have it too, and calling `applyPreset( parent, preset )` will apply it to all the children.
-- RenderController : Added `pathForID()`, `pathsForIDs()`, `idForPath()` and `idsForPaths()` methods. These make it possible to identify an object in the scene from a `uint id` AOV.
 
 Breaking Changes
 ----------------